+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 Changes:
 
-- fix: UNCOMMON color
+- fix: missing enchants not working on longswords
+- fix: some issues with the missing enchants caching, making it so a maxed item could have the list of missing enchants from other items.